exception handling

+++ db/db_user.py
@@ -1,3 +1,4 @@
+from fastapi import HTTPException, status
 from sqlalchemy.orm.session import Session
 
 from db.hash import Hash
@@ -22,10 +23,22 @@ def get_all_users(db: Session):
 
 
 def get_user(db: Session, id: int):
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f'User with id {id} not found'
+        )
+    return user
+
 
 def update_user(db: Session, id: int, request: UserBase):
     user = db.query(DbUser).filter(DbUser.id == id)
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f'User with id {id} not found'
+        )
     user.update(
         {
             DbUser.username: request.username,
@@ -36,8 +49,14 @@ def update_user(db: Session, id: int, request: UserBase):
     db.commit()
     return 'ok'
 
+
 def delete_user(db: Session, id: int):
     user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail=f'User with id {id} not found'
+        )
     db.delete(user)
     db.commit()
     return 'ok'
